Detach CLI functionality from example app code
